address most pr feedback
