+++ aggregator/src/main/scala/ai/chronon/aggregator/windowing/SawtoothMutationAggregator.scala
@@ -127,12 +127,13 @@ class SawtoothMutationAggregator(aggregations: Seq[Aggregation],
   }
 
   def updateIrTiled(ir: Array[Any], otherIr: TiledIr, queryTs: Long) = {
+    val otherIrTs = otherIr.ts
     var i: Int = 0
     while (i < windowedAggregator.length) {
       val window = windowMappings(i).aggregationPart.window
       val hopIndex = tailHopIndices(i)
       val irInWindow =
-        (otherIr.ts >= TsUtils.round(queryTs - window.millis, hopSizes(hopIndex)) && otherIr.ts < queryTs)
+        (otherIrTs >= TsUtils.round(queryTs - window.millis, hopSizes(hopIndex)) && otherIrTs < queryTs)
       if (window == null || irInWindow) {
         ir(i) = windowedAggregator(i).merge(ir(i), otherIr.ir(i))
       }
