+++ api/src/test/scala/ai/chronon/api/test/ExtensionsTest.scala
@@ -19,9 +19,8 @@ package ai.chronon.api.test
 import ai.chronon.api.{Accuracy, Builders, Constants, GroupBy}
 import org.junit.Test
 import ai.chronon.api.Extensions._
-import org.junit.Assert.{assertEquals, assertTrue}
-import org.mockito.Mockito
-import org.mockito.Mockito.{doReturn, mock, spy, when}
+import org.junit.Assert.{assertEquals, assertFalse, assertTrue}
+import org.mockito.Mockito.{spy, when}
 
 import scala.util.ScalaJavaConversions.JListOps
 import java.util.Arrays
@@ -75,9 +74,7 @@ class ExtensionsTest {
     val groupByMetadata = Builders.MetaData(name = "test")
     val groupBy = Builders.GroupBy(keyColumns = Seq("a", "c"), metaData = groupByMetadata)
     val joinPart = Builders.JoinPart(groupBy = groupBy)
-    val join = Builders.Join(
-      joinParts = Seq(joinPart),
-      skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
+    val join = Builders.Join(joinParts = Seq(joinPart), skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
     assertTrue(join.partSkewFilter(joinPart).nonEmpty)
     assertEquals("a NOT IN (b) OR c NOT IN (d)", join.partSkewFilter(joinPart).get)
   }
@@ -88,9 +85,7 @@ class ExtensionsTest {
     val groupBy = Builders.GroupBy(keyColumns = Seq("c"), metaData = groupByMetadata)
 
     val joinPart = Builders.JoinPart(groupBy = groupBy)
-    val join = Builders.Join(
-      joinParts = Seq(joinPart),
-      skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
+    val join = Builders.Join(joinParts = Seq(joinPart), skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
 
     assertTrue(join.partSkewFilter(joinPart).nonEmpty)
     assertEquals("c NOT IN (d)", join.partSkewFilter(joinPart).get)
@@ -102,9 +97,7 @@ class ExtensionsTest {
     val groupBy = Builders.GroupBy(keyColumns = Seq("x", "c"), metaData = groupByMetadata)
 
     val joinPart = Builders.JoinPart(groupBy = groupBy, keyMapping = Map("a" -> "x"))
-    val join = Builders.Join(
-      joinParts = Seq(joinPart),
-      skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
+    val join = Builders.Join(joinParts = Seq(joinPart), skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
 
     assertTrue(join.partSkewFilter(joinPart).nonEmpty)
     assertEquals("x NOT IN (b) OR c NOT IN (d)", join.partSkewFilter(joinPart).get)
@@ -116,9 +109,7 @@ class ExtensionsTest {
     val groupBy = Builders.GroupBy(keyColumns = Seq("non_existent"), metaData = groupByMetadata)
 
     val joinPart = Builders.JoinPart(groupBy = groupBy)
-    val join = Builders.Join(
-      joinParts = Seq(joinPart),
-      skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
+    val join = Builders.Join(joinParts = Seq(joinPart), skewKeys = Map("a" -> Seq("b"), "c" -> Seq("d")))
 
     assertTrue(join.partSkewFilter(joinPart).isEmpty)
   }
@@ -152,5 +143,21 @@ class ExtensionsTest {
     assertTrue(keys.contains(partitionColumn))
     assertTrue(keys.contains(Constants.TimeColumn))
     assertEquals(4, keys.size)
+  }
+
+  @Test
+  def testIsTilingEnabled(): Unit = {
+    def buildGroupByWithCustomJson(customJson: String = null): GroupBy =
+      Builders.GroupBy(
+        metaData = Builders.MetaData(name = "featureGroupName", customJson = customJson)
+      )
+
+    // customJson not set defaults to false
+    assertFalse(buildGroupByWithCustomJson().isTilingEnabled)
+    assertFalse(buildGroupByWithCustomJson("{}").isTilingEnabled)
+
+    assertTrue(buildGroupByWithCustomJson("{\"enable_tiling\": true}").isTilingEnabled)
+    assertFalse(buildGroupByWithCustomJson("{\"enable_tiling\": false}").isTilingEnabled)
+    assertFalse(buildGroupByWithCustomJson("{\"enable_tiling\": \"string instead of bool\"}").isTilingEnabled)
   }
 }
