+++ online/src/main/scala/ai/chronon/online/GroupByServingInfoParsed.scala
@@ -71,8 +71,9 @@ class GroupByServingInfoParsed(val groupByServingInfo: GroupByServingInfo, parti
   def outputCodec: AvroCodec = AvroCodec.of(outputAvroSchema)
 
   // Start tiling specific variables
+
   lazy val tiledCodec: TileCodec = new TileCodec(groupBy, valueChrononSchema.fields.map(sf => (sf.name, sf.fieldType)))
-  lazy val isTilingEnabled: Boolean = TileCodec.isTilingEnabled(groupBy)
+  lazy val isTilingEnabled: Boolean = groupByOps.isTilingEnabled
 
   // End tiling specific variables
